reorganize and remove models folder
