+++ mlpf/__init__.py
@@ -1,1 +1,2 @@
+from mlpf.model import MLPF
 from mlpf.utils import make_Rmaps
