--- models/__init__.py
@@ -1,3 +1,0 @@
-from models.fcn import FCN
-from models.mlpf import MLPF
-from models.particlenet import ParticleNet
