+++ particlenet/__init__.py
@@ -1,3 +1,5 @@
+from particlenet.model import ParticleNet
+
 from particlenet.dataset import TopTaggingDataset
 from particlenet.training import training_loop
 from particlenet.utils import load_data, make_roc, save_model
