+++ particlenet/run_inference.py
@@ -18,10 +18,7 @@ import torch_geometric
 from torch_geometric.data import Batch, Data
 from torch_geometric.loader import DataListLoader, DataLoader
 
-sys.path.insert(0, "..")
-from models import ParticleNet
-
-from particlenet import load_data, make_roc
+from particlenet import ParticleNet, load_data, make_roc
 
 plt.style.use(hep.style.CMS)
 plt.rcParams.update({"font.size": 20})
