+++ run_lrp_fcn.py
@@ -7,13 +7,46 @@ from glob import glob
 import matplotlib.pyplot as plt
 import numpy as np
 import torch
+import torch.nn as nn
 from sklearn.metrics import accuracy_score
 from torch.utils.data import DataLoader
 
 from explainer import LRP
-from models import FCN
 
 # this script builds a toy dataset, trains a simple FFN model on the dataset, and tests LRP
+
+
+class FCN(nn.Module):
+    """
+    Showcase an example of an fully connected network model, with a skip connection, that can be explained by LRP
+    """
+
+    def __init__(self, input_dim=3, hidden_dim=256, embedding_dim=40, output_dim=2):
+        super(FCN, self).__init__()
+
+        self.act = nn.ReLU
+
+        self.nn1 = nn.Sequential(
+            nn.Linear(input_dim, hidden_dim),
+            nn.BatchNorm1d(hidden_dim),
+            self.act(),
+            nn.Linear(hidden_dim, hidden_dim),
+            nn.BatchNorm1d(hidden_dim),
+            self.act(),
+            nn.Linear(hidden_dim, hidden_dim),
+            nn.BatchNorm1d(hidden_dim),
+            self.act(),
+            nn.Linear(hidden_dim, embedding_dim),
+        )
+        self.nn2 = nn.Sequential(
+            nn.Linear(input_dim + embedding_dim, hidden_dim),
+            self.act(),
+            nn.Linear(hidden_dim, output_dim),
+        )
+
+    def forward(self, X):
+        embedding = self.nn1(X)
+        return self.nn2(torch.cat([X, embedding], axis=1))
 
 
 def build_toy_dataset():
